services/ai-processing/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
pydantic-settings
cryptography
anthropic
openai
pgvector
numpy
PyJWT
